+++ app/src/main/res/layout/fragment_notifications.xml
@@ -11,7 +11,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@drawable/bg_main"
-        tools:context=".presentation.ui.main.NotificationsFragment">
+        tools:context=".presentation.ui.notifications.NotificationsFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv_notifications"
